MS-1507 Liberalise email validation
The email validation bundled with hof is overly restrictive.
This commit liberalises the validation to essentially accept anything
that has one @ sign and at least one dot in the domain portion of the
address.

A particular problem with the old code is that it wouldn't accept
apostrophes in the username portion; this commit fixes that.