MS-929 Add full-width behaviour
This commit lets you customise the width of the content container on a
per step basis. Previously, they were all hardcoded to two thirds of
page width, which is great most of the time but not helpful for the
final "check your answers" page.

This commit overrides the template partial defined in
hof-template-partials to allow a res.local.columnWidth to amend the
class of the content container. It also defines a new "full width"
behaviour in order to set that new local.

We also add this behaviour to the "check your answers" page of the nrm
app. The other apps don't seem to have a corresponding CYA page.