Use secondary button styling for "save and exit"
It's currently unclear which button to click to proceed to the next
page. This commit moves the "save and exit" button to be after the
"continue" button, and colours it grey instead of green.

For more information on secondary buttons, see the GOV.UK Design System:
https://design-system.service.gov.uk/components/button/#secondary-buttons

The concept of secondary buttons did not exist in GOV.UK Elements, which
hof continues to use. Therefore, this commit also adds some SCSS to
add a .button-secondary class.